Removed Impl namespace and rearranged files
