+++ DNN Platform/Modules/TelerikRemovalLibrary/Dnn.Modules.TelerikRemovalLibrary.csproj
@@ -53,37 +53,37 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Impl\ClearCacheStep.cs" />
-    <Compile Include="Impl\ExecuteSqlStep.cs" />
-    <Compile Include="Impl\FileSystemProvider.cs" />
-    <Compile Include="Impl\IClearCacheStep.cs" />
-    <Compile Include="Impl\IDataCache.cs" />
-    <Compile Include="Impl\IDataProvider.cs" />
-    <Compile Include="Impl\IDesktopModuleController.cs" />
-    <Compile Include="Impl\IExecuteSqlStep.cs" />
-    <Compile Include="Impl\IFileSystemProvider.cs" />
-    <Compile Include="Impl\IInstallAvailablePackageStep.cs" />
-    <Compile Include="Impl\IModuleDefinitionController.cs" />
-    <Compile Include="Impl\InstallAvailablePackageStep.cs" />
-    <Compile Include="Impl\InternalDataCache.cs" />
-    <Compile Include="Impl\InternalDataProvider.cs" />
-    <Compile Include="Impl\InternalDesktopModuleController.cs" />
-    <Compile Include="Impl\InternalModuleDefinitionController.cs" />
-    <Compile Include="Impl\INullStep.cs" />
-    <Compile Include="Impl\IReplacePortalTabModuleStep.cs" />
-    <Compile Include="Impl\IReplaceTabModuleStep.cs" />
-    <Compile Include="Impl\IStep.cs" />
-    <Compile Include="Impl\IStepArray.cs" />
-    <Compile Include="Impl\NullStep.cs" />
-    <Compile Include="Impl\ReplacePortalTabModuleStep.cs" />
-    <Compile Include="Impl\ReplaceTabModuleStep.cs" />
-    <Compile Include="Impl\RequiredAttribute.cs" />
-    <Compile Include="Impl\StepBase.cs" />
-    <Compile Include="Impl\TelerikUninstaller.cs" />
     <Compile Include="ITelerikUninstaller.cs" />
-    <Compile Include="UninstallSummaryItem.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Shims\FileSystemProvider.cs" />
+    <Compile Include="Shims\IDataCache.cs" />
+    <Compile Include="Shims\IDataProvider.cs" />
+    <Compile Include="Shims\IDesktopModuleController.cs" />
+    <Compile Include="Shims\IFileSystemProvider.cs" />
+    <Compile Include="Shims\IModuleDefinitionController.cs" />
+    <Compile Include="Shims\InternalDataCache.cs" />
+    <Compile Include="Shims\InternalDataProvider.cs" />
+    <Compile Include="Shims\InternalDesktopModuleController.cs" />
+    <Compile Include="Shims\InternalModuleDefinitionController.cs" />
     <Compile Include="Startup.cs" />
+    <Compile Include="Steps\ClearCacheStep.cs" />
+    <Compile Include="Steps\ExecuteSqlStep.cs" />
+    <Compile Include="Steps\IClearCacheStep.cs" />
+    <Compile Include="Steps\IExecuteSqlStep.cs" />
+    <Compile Include="Steps\IInstallAvailablePackageStep.cs" />
+    <Compile Include="Steps\InstallAvailablePackageStep.cs" />
+    <Compile Include="Steps\INullStep.cs" />
+    <Compile Include="Steps\IReplacePortalTabModuleStep.cs" />
+    <Compile Include="Steps\IReplaceTabModuleStep.cs" />
+    <Compile Include="Steps\IStep.cs" />
+    <Compile Include="Steps\IStepArray.cs" />
+    <Compile Include="Steps\NullStep.cs" />
+    <Compile Include="Steps\ReplacePortalTabModuleStep.cs" />
+    <Compile Include="Steps\ReplaceTabModuleStep.cs" />
+    <Compile Include="Steps\RequiredAttribute.cs" />
+    <Compile Include="Steps\StepBase.cs" />
+    <Compile Include="TelerikUninstaller.cs" />
+    <Compile Include="UninstallSummaryItem.cs" />
   </ItemGroup>
   <ItemGroup>
     <AdditionalFiles Include="..\..\..\stylecop.json">
