+++ DNN Platform/Modules/TelerikRemovalLibrary/Shims/IDataCache.cs
@@ -1,0 +1,19 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information
+
+namespace Dnn.Modules.TelerikRemovalLibrary
+{
+    using DotNetNuke.Common.Utilities;
+
+    /// <summary>
+    /// Abstraction of the <see cref="DataCache"/> class to enable DI and unit testing.
+    /// </summary>
+    internal interface IDataCache
+    {
+        /// <summary>
+        /// Clears the cache.
+        /// </summary>
+        void ClearCache();
+    }
+}
