+++ DNN Platform/Modules/TelerikRemovalLibrary/Shims/InternalDesktopModuleController.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information
 
-namespace Dnn.Modules.TelerikRemovalLibrary.Impl
+namespace Dnn.Modules.TelerikRemovalLibrary
 {
     using DotNetNuke.Entities.Modules;
